+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Fixed
+- MdnsServer::SetHostname restarting the server even when the hostname did not change.
+
 ## [1.2.3] - 2026-08-26
 ### Fixed
 - mdns_server example and test server not stopping the server task before destruction.